+++ examples/pipe.js
@@ -1,4 +1,4 @@
-const { getPath, renderFn } = require('../')
+const { pathGet, renderFn } = require('../')
 
 function compose(fnArr, initialValue) {
   return fnArr.reduce((input, fn) => fn(input), initialValue)
@@ -17,7 +17,7 @@ function pipe(template, scope) {
   function applyPipe(path) {
     // We are going to let this throw in case it cannot parse it
     const [initialValuePath, ...pipeNames] = path.split('|')
-    const value = getPath(scope, initialValuePath)
+    const value = pathGet(scope, initialValuePath)
     if (pipeNames.length) {
       const fnArr = pipeNames.map((pipeName) => pipes[pipeName.trim()])
       return compose(fnArr, value)